Create codeql-analysis.yml (#562)
* Create codeql-analysis.yml

* MAINT: Replace readme

* FIX: Build

* FIX: Better

* FIX: Better

* FIX: Compact

* FIX: NumPy

* FIX: Three alerts

+++ OpenMEEGMaths/src/matrix.cpp
@@ -37,7 +37,7 @@ namespace OpenMEEG {
     }
 
     Matrix::Matrix(const Vector& v,const Dimension M,const Dimension N): LinOp(M,N,FULL,2) {
-        om_assert(M*N==v.size());
+        om_assert(static_cast<unsigned long>(M)*static_cast<unsigned long>(N)==v.size());
         value = v.value;
     }
